% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-blacklisthighmap-v2.R,
%   R/preprocessing-blacklisthighmap.R
\name{blacklisthighmap}
\alias{blacklisthighmap}
\title{Blacklist High Mappability Regions in Genomic Data}
\usage{
blacklisthighmap(maptrackpath, blacklistshpath, exptabpath,
   nbcputrans, allwindowsbed, windsize, genomename, saveobjectpath = NA,
   tmpfold = "./tmp", reload = FALSE, showtime = FALSE, showmemory = FALSE,
   verbose = TRUE)

blacklisthighmap(maptrackpath, blacklistshpath, exptabpath, nbcputrans,
 allwindowsbed, windsize, saveobjectpath = NA, reload = FALSE,
 showtime = FALSE, verbose = TRUE)
}
\arguments{
\item{maptrackpath}{Character string. Path to the mappability track file.}

\item{blacklistshpath}{Character string. Path to the blacklist regions file.}

\item{exptabpath}{Path to the experiment table file containing a table with
columns named 'condition', 'replicate', 'strand', and 'path'.}

\item{nbcputrans}{Number of CPU cores to use for transcript-level operations.}

\item{allwindowsbed}{Data frame. BED-formatted data frame obtained with the
function makewindows.}

\item{windsize}{Window size for splitting transcripts into intervals.}

\item{saveobjectpath}{Path to save intermediate R objects. Default is `NA`
and R objects are not saved.}

\item{reload}{Logical. If `TRUE`, reloads existing saved objects to avoid
recomputation. Default is `FALSE`. If the function failed during object
saving, make sure to delete the corresponding object.}

\item{showtime}{Logical. Whether to display timing information.}

\item{verbose}{Logical. Whether to display detailed progress messages.}

\item{blacklistpath}{Character string. Path to the blacklist regions file.}

\item{genomename}{Character string. A valid UCSC genome name. It is used to
retrieve chromosome metadata, such as names and lengths.}

\item{tmpfold}{A character string specifying the temporary folder for saving
output files. The temporary files contain the scores for each bedgraph on
each chromosome.}

\item{showmemory}{A logical value indicating whether to display memory usage 
during processing.}
}
\value{
This function does not return a value directly. It saves
intermediate results to `tmpfold`. These intermediates files are then
combined by the function 'createtablescores'.

A list of data frames where each entry corresponds to the processed
 scores for an experiment. Scores outside high-mappability regions or in
 blacklisted regions are set to `NA`.
}
\description{
This function processes genomic data to remove scores that fall within 
blacklisted regions or have low mappability, and computes weighted means for
overlapping windows. The process ensures the integrity of genomic scores by
focusing on high mappability regions and excluding blacklisted intervals.

This function processes bedgraph files by filtering out scores overlapping
blacklisted regions and retaining those in high-mappability regions. It
computes weighted means for scores within overlapping windows and supports
parallel processing for efficiency.
}
\details{
The `blacklisthighmap` function iterates through chromosomes, processes
genomic scores by removing those overlapping with blacklisted regions, and
ensures that scores within windows are computed using a weighted mean when
overlaps occur. The function uses parallel processing for efficiency and
supports saving (saveobjectpath) and reloading (reload) intermediate results
to optimize workflow.

The main steps include:
- Reading and processing bedGraph values.
- Removing scores overlapping with blacklisted or low mappability regions.
- Computing weighted means for overlapping scores in genomic windows.
- Saving the processed results to specified path (tmpfold).

The function involves the following steps:
1. Reading and converting the blacklist, mappability track, and annotation
 windows into tibbles.
2. Reading experiment metadata to identify bedgraph files.
3. For each bedgraph file:
   - Extracting scores based on strand information.
   - Filtering scores overlapping blacklisted regions or outside
 high-mappability intervals.
   - Computing weighted means for overlapping windows.
4. Combining processed scores into a single data frame for each experiment.

The function can save intermediate objects to disk and reload them for
efficiency in repeated runs.
}
\examples{
# Define paths to required files
maptrackpath <- "path/to/maptrack.bed"
blacklistshpath <- "path/to/blacklist.bed"
exptabpath <- "path/to/experiments.csv"
allwindowsbed <- data.frame(...)

# Run the function
results <- blacklisthighmap(
    maptrackpath = maptrackpath,
    blacklistshpath = blacklistshpath,
    exptabpath = exptabpath,
    nbcputrans = 4,
    allwindowsbed = allwindowsbed,
    windsize = 200,
    genomename = "hg38",
    saveobjectpath = "output/",
    tmpfold = "./tmp",
    reload = FALSE,
    showtime = TRUE,
    showmemory = FALSE,
    verbose = TRUE)

# Define paths to required files
maptrackpath <- "path/to/maptrack.bed"
blacklistshpath <- "path/to/blacklist.bed"
exptabpath <- "path/to/experiments.csv"
allwindowsbed <- data.frame(...)

# Run the function
results <- blacklisthighmap(
    maptrackpath = maptrackpath,
    blacklistshpath = blacklistshpath,
    exptabpath = exptabpath,
    nbcputrans = 4,
    allwindowsbed = allwindowsbed,
    windsize = 100,
    saveobjectpath = "output/",
    reload = FALSE,
    showtime = TRUE,
    verbose = TRUE
)

}
\seealso{
[createtablescores][makewindows]

[makewindows]
}
