% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-createtablescores-v2.R,
%   R/preprocessing-createtablescores.R
\name{createtablescores}
\alias{createtablescores}
\title{Create a Unified Table of Scores}
\usage{
createtablescores(tmpfold, exptabpath, showmemory = FALSE, showtime = TRUE,
  savefinaltable = TRUE, finaltabpath = "./", finaltabname = "anno.tsv",
 verbose)

createtablescores(bedgraphlistwmean, nbcpubg, exptabpath,
 saveobjectpath = NA, reload = FALSE, showtime = TRUE, verbose = TRUE)
}
\arguments{
\item{bedgraphlistwmean}{A list for each bedgraph with values on transcript
windows. It was obtained with the function blacklisthighmap.}

\item{nbcpubg}{Number of CPU cores to use for bedgraph-level operations.}

\item{exptabpath}{Path to the experiment table file containing a table with
columns named 'condition', 'replicate', 'strand', and 'path'.}

\item{saveobjectpath}{Path to save intermediate R objects. Default is `NA`
and R objects are not saved.}

\item{reload}{Logical. If `TRUE`, reloads existing saved objects to avoid
recomputation. Default is `FALSE`. If the function failed during object
saving, make sure to delete the corresponding object.}

\item{showtime}{Logical. If `TRUE`, displays timing information. Default is
`FALSE`.}

\item{verbose}{Logical. If `TRUE`, provides detailed messages during
execution. Default is `TRUE`.}

\item{tmpfold}{A string specifying the temporary folder containing the
score files created with the function 'blacklisthighmap'.}

\item{showmemory}{Logical; if `TRUE`, memory usage is printed during
processing. Default is `FALSE`.}

\item{savefinaltable}{Logical; if `TRUE`, the resulting table is saved to
disk. Default is `TRUE`.}

\item{finaltabpath}{A string specifying the directory where the final table
should be saved. Default is `"./"`.}

\item{finaltabname}{A string specifying the name of the final table file.
Default is `"anno.tsv"`.}
}
\value{
A data frame containing the unified table of scores.

A tibble containing the final merged and processed table with scores
and experiment details.
}
\description{
This function processes and combines table scores of each bedgraph and each
chromosome stored in the temporary folder into a unified table.

This function processes the scores of bedgraph files retrieved on transcripts
to create a merged table of scores. It performs tasks such as adding unique
row IDs, joining bedgraph data, reordering and renaming columns, and
preparing a final table for further analysis.
}
\details{
This function first merges files belonging to the same experiment and
direction. These files are combined into a single table providing two columns
per experiment. The first gives the name of the experiment and the second the
scores. The resulting table also includes annotations for each transcript.

The function performs the following steps:
1. Reads experiment details from the provided exptabpath.
2. Adds a unique row ID to each bedgraph data frame based on transcript,
 gene, strand, and window.
3. Merges all bedgraph data frames into a single data frame using the unique
 row ID.
4. Reorders and renames columns for consistency and clarity.
5. Adds experiment-related columns based on the provided experiment table.
6. Optionally saves the resulting table to the specified path.
}
\examples{
# Example usage:
tmpfold <- "path/to/temp/folder"
exptabpath <- "path/to/experiment_table.csv"
finaltab <- createtablescores(tmpfold = tmpfold, exptabpath = exptabpath,
  showmemory = TRUE, showtime = TRUE, savefinaltable = TRUE,
  finaltabpath = "./results", finaltabname = "final_scores.tsv",
  verbose = TRUE)

# Example usage of createtablescores function:
bedgraphlist <- list(bedgraph1, bedgraph2, bedgraph3)
exptabpath <- "path/to/experiment_table.csv"
output_path <- "path/to/save_directory"
final_table <- createtablescores(bedgraphlist, nbcpubg = 4, exptabpath,
 saveobjectpath = output_path, reload = FALSE, showtime = TRUE,
 verbose = TRUE)

}
\seealso{
[blacklisthighmap]

[blacklisthighmap]
}
