% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{joinfiles}
\alias{joinfiles}
\title{Join Bedgraph Files for Protein-Coding and lncRNA Data}
\usage{
joinfiles(workingdir = ".", window = 200, bgpattern = "*.bg",
  protscoredir = "protein_coding_score", lncscoredir = "lncRNA_score",
  outtsv = "dTAG_Cugusi_stranded_20230810.tsv", verbose = TRUE)
}
\arguments{
\item{workingdir}{The directory containing bedgraph files. Defaults to the
current working directory (`"."`).}

\item{window}{The window size used for joining the score files. Defaults to
200.}

\item{bgpattern}{A file pattern to identify bedgraph files. Defaults to
`"*.bg"`.}

\item{protscoredir}{Directory containing the protein-coding score files.
Defaults to `"protein_coding_score"`.}

\item{lncscoredir}{Directory containing the lncRNA score files. Defaults to
`"lncRNA_score"`.}

\item{outtsv}{The output TSV filename where the merged data will be saved.
Defaults to `"dTAG_Cugusi_stranded_20230810.tsv"`.}

\item{nbcpu}{An integer specifying the number of CPU cores to use for
parallel computation. Default is \code{1}.}

\item{verbose}{Logical flag to enable verbose output during the function
execution. Defaults to `TRUE`.}
}
\value{
The data.frame with the complete set of annotations and scores.
}
\description{
The `joinfiles` function processes bedgraph files located in the specified
working directory, joins data related to protein-coding and lncRNA
annotations, and outputs a combined result as a TSV file. The function
retrieves bedgraph files matching a specific pattern, processes the files to
create windows-based summaries, and merges annotations for protein-coding and
lncRNA biotypes. The resulting data is written to an output file.
}
\examples{
\dontrun{
  joinfiles(workingdir = "data", window = 100, bgpattern = "*.bedgraph",
    protscoredir = "prot_scores", lncscoredir = "lnc_scores",
    outtsv = "results.tsv")
}

}
