% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meandifference.R
\name{meandifference}
\alias{meandifference}
\title{Compute Mean and Differences of Scores for Each Condition}
\usage{
meandifference(resultsecdf, expdf, nbwindows, showtime = FALSE,
verbose = TRUE)
}
\arguments{
\item{resultsecdf}{A data frame containing ECDF results for each transcript
and condition (see genesECDF).}

\item{expdf}{A data frame containing experimental information including a
\code{condition} column.}

\item{nbwindows}{An integer representing the number of windows (or segments)
in each transcript.}

\item{showtime}{A logical value indicating if the duration of the function
processing should be indicated before ending. Defaults to
\code{FALSE}.}

\item{verbose}{A logical flag indicating whether to print progress messages.
Defaults to \code{TRUE}.}
}
\value{
A data frame that contains, for each condition:
\itemize{
  \item Mean values for the "value" and "Fx" columns (e.g.,
   \code{mean_value_ctrl}, \code{mean_Fx_ctrl}).
  \item Differences between the \code{Fx} column and coordinate ratios
   (e.g., \code{diff_Fx_ctrl}).
}
}
\description{
This function calculates the mean values, mean Fx (ECDF) and ECDF differences
(Fx) for expression data, across different experimental conditions.
}
\examples{
# Assuming resultsecdf is a data frame with ECDF results and expdf contains
# conditions:
# result <- meandifference(resultsecdf, expdf, nbwindows = 200,
# verbose = TRUE)

}
