% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-createtablescores.R
\name{createtablescores}
\alias{createtablescores}
\title{Create a Unified Table of Scores}
\usage{
createtablescores(tmpfold, exptabpath, showmemory = FALSE, showtime = TRUE,
  savefinaltable = TRUE, finaltabpath = "./", finaltabname = "anno.tsv",
 verbose)
}
\arguments{
\item{tmpfold}{A string specifying the temporary folder containing the
score files created with the function 'blacklisthighmap'.}

\item{exptabpath}{Path to the experiment table file containing a table with
columns named 'condition', 'replicate', 'strand', and 'path'.}

\item{showmemory}{Logical; if `TRUE`, memory usage is printed during
processing. Default is `FALSE`.}

\item{showtime}{Logical; if `TRUE`, the execution time of the function is
printed. Default is `TRUE`.}

\item{savefinaltable}{Logical; if `TRUE`, the resulting table is saved to
disk. Default is `TRUE`.}

\item{finaltabpath}{A string specifying the directory where the final table
should be saved. Default is `"./"`.}

\item{finaltabname}{A string specifying the name of the final table file.
Default is `"anno.tsv"`.}

\item{verbose}{Logical; if `TRUE`, detailed messages are printed during
execution.}
}
\value{
A data frame containing the unified table of scores.
}
\description{
This function processes and combines table scores of each bedgraph and each
chromosome stored in the temporary folder into a unified table.
}
\details{
This function first merges files belonging to the same experiment and
direction. These files are combined into a single table providing two columns
per experiment. The first gives the name of the experiment and the second the
scores. The resulting table also includes annotations for each transcript.
}
\examples{
# Example usage:
tmpfold <- "path/to/temp/folder"
exptabpath <- "path/to/experiment_table.csv"
finaltab <- createtablescores(tmpfold = tmpfold, exptabpath = exptabpath,
  showmemory = TRUE, showtime = TRUE, savefinaltable = TRUE,
  finaltabpath = "./results", finaltabname = "final_scores.tsv",
  verbose = TRUE)

}
\seealso{
[blacklisthighmap]
}
